+++ examples/datasets.py
@@ -1,8 +1,8 @@
 '''
-Example of the basic operations with ``pyroomacoustics.datasets.CorpusBase``
-and ``pyroomacoustics.datasets.SampleBase`` classes
+Example of the basic operations with ``pyroomacoustics.datasets.Dataset``
+and ``pyroomacoustics.datasets.Sample`` classes
 '''
-from pyroomacoustics.datasets import SampleBase, CorpusBase
+from pyroomacoustics.datasets import Sample, Dataset
 
 # Prepare a few artificial samples
 samples = [
@@ -32,24 +32,31 @@ samples = [
     },
     ]
 
-corpus = CorpusBase()
+dataset = Dataset()
 for s in samples:
-    new_sample = SampleBase(s['data'], **s['metadata'])
-    corpus.add_sample(new_sample)
+    new_sample = Sample(s['data'], **s['metadata'])
+    dataset.add_sample(new_sample)
 
-# Then, it possible to display summary info about the corpus
-print(corpus)
+# Then, it possible to display summary info about the dataset
+print(dataset)
 print()
 
-# The number of samples in the corpus is given by ``len``
-print('Number of samples:', len(corpus))
+# The number of samples in the dataset is given by ``len``
+print('Number of samples:', len(dataset))
 print()
 
 # And we can access samples with the slice operator
 print('Sample #2:')
-print(corpus[2])
+print(dataset[2])
 print()
 
-# We can obtain a new corpus with only male subject
-corpus_male_only = corpus.filter(sex='male')
-print(corpus_male_only)
+# We can obtain a new dataset with only male subject
+dataset_male_only = dataset.filter(sex='male')
+print('Dataset with only males')
+print(dataset_male_only)
+print()
+
+# Only retain speakers above 40 years old
+dataset_older = dataset.filter(age=lambda a : a > 40)
+print('Dataset with only above 40 yo')
+print(dataset_older)