+++ CHANGELOG.rst
@@ -11,6 +11,28 @@ adheres to `Semantic Versioning <http://semver.org/spec/v2.0.0.html>`_.
 `Unreleased`_
 -------------
 
+Added
+~~~~~
+
+- Base classes for Dataset and Sample in ``pyroomacoustics.datasets``
+- Methods to filter datasets according to the metadata of samples
+- Deprecation warning for the TimitCorpus interface
+
+Changed
+~~~~~~~
+
+- CMUArcticDatabase basedir is now the top directory where CMU_ARCTIC database
+  should be saved. Not the directory above as it previously was.
+
+
+`0.1.15`_ - 2018-06-21
+----------------------
+
+Bugfix
+~~~~~~
+
+- Added ``pyroomacoustics.datasets`` to list of sub-packages in ``setup.py``
+
 `0.1.14`_ - 2018-06-20
 ----------------------
 
@@ -26,5 +48,6 @@ Changed
 -  Moved TIMIT corpus wrapper from ``pyroomacoustics.recognition`` module to sub-package
    ``pyroomacoustics.datasets.timit``
 
-.. _Unreleased: https://github.com/olivierlacan/keep-a-changelog/compare/v0.1.14...HEAD
+.. _Unreleased: https://github.com/olivierlacan/keep-a-changelog/compare/v0.1.15...HEAD
+.. _0.1.15: https://github.com/olivierlacan/keep-a-changelog/compare/v0.1.14...v0.1.15
 .. _0.1.14: https://github.com/olivierlacan/keep-a-changelog/compare/v0.1.13...v0.1.14